Created custom UIView for the format view
