+++ MacroPepelelipa/MacroPepelelipa/View/Markup/MarkupContainerView.swift
@@ -8,30 +8,8 @@
 
 import UIKit
 
-enum FormatSelector {
-    case italic
-    case bold
-    case highlight
-}
-
-enum ColorSelector {
-    case black
-    case green
-    case red
-}
-
-enum FontSelector: String {
-    case merriweather = "Merriweather"
-    case openSans = "OpenSans"
-    case dancingScript = "Dancing"
-}
-
-internal class MarkupContainerView: UIView, TextEditingDelegateObserver {
+internal class MarkupContainerView: MarkupFormatView, TextEditingDelegateObserver {
     
-    private weak var textView: MarkupTextView?
-    private weak var viewController: NotesViewController?
-    public weak var delegate: MarkupFormatViewDelegate?
-        
     private lazy var backgroundView: UIView = {
         let bckView = UIView(frame: .zero)
         bckView.backgroundColor = .white
@@ -58,77 +36,12 @@ internal class MarkupContainerView: UIView, TextEditingDelegateObserver {
         return fmtLabel
     }()
     
-    private lazy var colorSelector: [ColorSelector: MarkupToggleButton] = {
-        var buttons = [ColorSelector: MarkupToggleButton]()
-        var buttonColors: [UIColor] = [UIColor.bodyColor ?? .black, UIColor.notebookColors[4], UIColor.notebookColors[14]]
-        
-        buttonColors.forEach { (color) in
-            var newButton = MarkupToggleButton(frame: .zero, color: color)
-            newButton.translatesAutoresizingMaskIntoConstraints = false
-            newButton.addTarget(delegate, action: #selector(delegate?.changeTextColor), for: .touchDown)
-            
-            if color == buttonColors[0] {
-                buttons[.black] = newButton
-            } else if color == buttonColors[1] {
-                buttons[.green] = newButton
-            } else {
-                buttons[.red] = newButton
-            }
-        }
-        
-        return buttons
-    }()
-    
-    private lazy var formatSelector: [FormatSelector: MarkupToggleButton] = {
-        var buttons = [FormatSelector: MarkupToggleButton]()
-        var imageNames: [FormatSelector: String] = [.italic: "italic", .bold: "bold", .highlight: "pencil.tip"]
-        
-        for (key, imageName) in imageNames {
-            var newButton = createButton(
-                normalStateImage: UIImage(systemName: imageName),
-                titleLabel: nil
-            )
-            newButton.translatesAutoresizingMaskIntoConstraints = false
-            
-            if key == .italic {
-                newButton.addTarget(delegate, action: #selector(delegate?.makeTextItalic), for: .touchDown)
-            } else if key == .bold {
-                newButton.addTarget(delegate, action: #selector(delegate?.makeTextBold), for: .touchDown)
-            } else {
-                newButton.addTarget(delegate, action: #selector(delegate?.highlightText), for: .touchUpInside)
-            }
-            
-            buttons[key] = newButton
-        }
-        
-        return buttons
-    }()
-    
-    private lazy var fontSelector: [FontSelector: MarkupToggleButton] = {
-        var buttons = [FontSelector: MarkupToggleButton]()
-        let systemFont = UIFont.systemFont(ofSize: UIFont.labelFontSize)
-        var fonts: [UIFont: FontSelector] = [UIFont.merriweather ?? systemFont: .merriweather, UIFont.openSans ?? systemFont: .openSans, UIFont.dancingScript ?? systemFont: .dancingScript]
-        
-        for font in fonts {
-            var newButton = createButton(
-                normalStateImage: nil,
-                titleLabel: font.value.rawValue,
-                font: font.key
-            )
-            newButton.translatesAutoresizingMaskIntoConstraints = false
-            newButton.addTarget(delegate, action: #selector(delegate?.changeTextFont), for: .touchDown)
-            buttons[font.value] = newButton
-        }
-        
-        return buttons
-    }()
-    
     init(frame: CGRect, owner: MarkupTextView, delegate: MarkupFormatViewDelegate?, viewController: NotesViewController) {
+        super.init(frame: frame)
         self.textView = owner
         self.delegate = delegate
         self.viewController = viewController
         
-        super.init(frame: frame)
         self.backgroundColor = UIColor.backgroundColor
         self.addSubview(backgroundView)
         
@@ -191,6 +104,8 @@ internal class MarkupContainerView: UIView, TextEditingDelegateObserver {
      This method sets the constraints for the inner elements of the container view.
      */
     public func createConstraints() {
+        backgroundView.addSubview(dismissButton)
+        backgroundView.addSubview(formatLabel)
         backgroundView.layer.zPosition = -1
         
         NSLayoutConstraint.activate([
@@ -331,53 +246,4 @@ internal class MarkupContainerView: UIView, TextEditingDelegateObserver {
         updateSelectors()
     }
     
-    /**
-     This public methos updates the selectors appearence based on the text style.
-     */
-    public func updateSelectors() {
-        guard let textView = self.textView else {
-            return
-        }
-        
-        formatSelector[.italic]?.isSelected = textView.checkTrait(.traitItalic)
-        formatSelector[.bold]?.isSelected = textView.checkTrait(.traitBold)
-        formatSelector[.highlight]?.isSelected = textView.checkBackground()
-
-        for (_, button) in formatSelector {
-            button.setTintColor()
-        }
-        
-        let textcolor = textView.getTextColor()
-        for (_, button) in colorSelector {
-            button.isSelected = (textcolor == button.backgroundColor)
-        }
-        
-        let textFont = textView.getTextFont()
-        for (_, button) in fontSelector {
-            button.isSelected = (textFont.familyName == button.titleLabel?.font.familyName)
-        }
-    }
-    
-    /**
-     This methos updates the color selectors by comparing each one with the sender button.
-     
-     - Parameter sender: The MarkupToggleButton that was last selected.
-     */
-    public func updateColorSelectors(sender: MarkupToggleButton) {
-        for (_, button) in colorSelector where button != sender {
-            button.isSelected = false
-        }
-    }
-    
-    /**
-     This methos updates the font selectors by comparing each one with the sender button.
-     
-     - Parameter sender: The MarkupToggleButton that was last selected.
-     */
-    public func updateFontSelectors(sender: MarkupToggleButton) {
-        for (_, button) in fontSelector where button != sender {
-            button.isSelected = false
-            button.setTintColor()
-        }
-    }
 }
